requirements.txt
# Flask Core
Flask==3.0.0
python-dotenv==1.0.0

# Rate Limiting
Flask-Limiter==3.5.0

# LLM Clients
ollama==0.1.0
requests==2.31.0

# OCR
pytesseract==0.3.10
Pillow==10.1.0

# Data Validation
pydantic==2.5.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-xdist==3.5.0
pytest-mock==3.12.0

# Utilities
python-dateutil==2.8.2
